+++ backend/src/main.py
@@ -1,0 +1,7 @@
+
+import asyncio
+import zoom
+import scribe
+
+asyncio.run(zoom.initialize())
+scribe.deliver()
